requirements.txt
openai
streamlit
python-dotenv
numpy
faiss-cpu
langchain
tiktoken
PyMuPDF
python-docx
Pillow
rank_bm25
sudachipy
sudachidict_core